requirements.txt
contourpy==1.2.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.48.1
kiwisolver==1.4.5
matplotlib==3.8.2
mplcursors==0.5.3
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
tk==0.1.0
